Change Torrent.(Size|Peers|Seeds) to int64

+++ bittorrent/torrent.go
@@ -20,9 +20,9 @@ type Torrent struct {
 	InfoHash  string   `json:"info_hash"`
 	Name      string   `json:"name"`
 	Trackers  []string `json:"trackers"`
-	Size      int      `json:"size"`
-	Seeds     int      `json:"seeds"`
-	Peers     int      `json:"peers"`
+	Size      int64    `json:"size"`
+	Seeds     int64    `json:"seeds"`
+	Peers     int64    `json:"peers"`
 	IsPrivate bool     `json:"is_private"`
 
 	Resolution  int    `json:"resolution"`
